added posterior object to new module
